Move confirmation email stuff to "appointments"
