+++ app/appointments/pages/schedule/[uid]/[slug].tsx
@@ -1,13 +1,13 @@
 import AvailableTimeSlotsSelection from "app/appointments/components/availableTimeSlotsSelection"
 import getMeeting from "app/appointments/queries/getMeeting"
 import getConnectedCalendars from "app/appointments/queries/getConnectedCalendars"
-import { BlitzPage, useQuery, useParam, invoke, useMutation } from "blitz"
+import { BlitzPage, useQuery, useParam, useMutation } from "blitz"
 import React, { Suspense, useState } from "react"
 import { DatePickerCalendar } from "react-nice-dates"
 import "react-nice-dates/build/style.css"
 import { enUS } from "date-fns/locale"
 import getTimeSlots from "app/appointments/queries/getTimeSlots"
-import sendConfirmationMailMutation from "app/components/createEmail/mutations/sendConfirmationMail"
+import sendConfirmationMailMutation from "app/appointments/mutations/sendConfirmationMail"
 import Card from "react-bootstrap/Card"
 import Row from "react-bootstrap/Row"
 import Col from "react-bootstrap/Col"
